test: added tests for updating, assigning and completing listeners
